Rename wrapper files for vendored libraries with 'vendored_' prefix
