--- lib/rubygems/net/http.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../vendor/net-http/lib/net/http"
