+++ lib/rubygems/vendored_net_http.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "vendor/net-http/lib/net/http"
