+++ lib/rubygems/vendored_optparse.rb
