+++ lib/rubygems/vendored_timeout.rb
