+++ lib/rubygems/vendored_tsort.rb
